Split utility-hooks into two files
