+++ gui/src/renderer/components/cell/Section.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { colors } from '../../../config.json';
 import { useAppContext } from '../../context';
 import { useHistory } from '../../lib/history';
-import { useBoolean, useEffectEvent } from '../../lib/utilityHooks';
+import { useBoolean, useEffectEvent } from '../../lib/utility-hooks';
 import Accordion from '../Accordion';
 import ChevronButton from '../ChevronButton';
 import { buttonText, openSans, sourceSansPro } from '../common-styles';
